replace all int IDs with Longs
nicer to be consistent, than having to remember which are int, which are long

+++ app/models/APIUser.scala
@@ -7,6 +7,6 @@ class APIUser(
                var apikey: String,
                var email: String,
                var role: Int // TODO this should be enum
-             ) extends KeyedEntity[Int] {
-  val id: Int = 0
+             ) extends KeyedEntity[Long] {
+  val id: Long = 0
 }
