+++ app/v1/pickee/PickeeRouter.scala
@@ -10,7 +10,7 @@ import play.api.routing.sird._
 class PickeeRouter @Inject()(controller: PickeeController) extends SimpleRouter {
   val prefix = "/v1/pickees"
 
-  def link(id: Int): String = {
+  def link(id: Long): String = {
     import com.netaporter.uri.dsl._
     val url = prefix / id.toString
     url.toString()
